Add files via upload
The automobile industry is a critical sector in the global economy, encompassing a wide range of businesses involved in the design, development, manufacturing, marketing, and selling of motor vehicles. This analysis report aims to provide a comprehensive overview of the current performance and trends within a specific automobile company. Utilizing Power BI, we have visualized key data metrics to deliver actionable insights that can aid in strategic decision-making.